Avoid waiting in exec_background(), closes #6
Instead, we have a pipe that is set to FD_CLOEXEC in the child,
so if there is nothing to read in the parent, then we know that
the exec was successful. If it was not, then we write the error
code in the pipe, so the parent can read it out.